+++ R/lazy_bootstrap.R
@@ -8,16 +8,17 @@ globalVariables("y")
 #' @param ci Confidence interval coverage, Default: `0.95`.
 #' @param ... Additional arguments.
 #' @return A `data.frame`.
-#' @rdname pmc_srmr
+#' @rdname pmc
 #' @export
 pmc_srmr <- function(x, ..., reps = 20, ci = .95){
-  UseMethod("pmc_srmr", x)
+  .Deprecated("pmc")
+  cl <- match.call()
+  cl[[1]] <- quote(pmc)
+  eval.parent(cl)
 }
 
 #' @importFrom stats quantile
-#' @method pmc_srmr mixture_list
-#' @export
-pmc_srmr.mixture_list <- function(x, ..., reps = 100, ci = .95){
+pmc_srmr_mixture_list <- function(x, ..., reps = 100, ci = .95){
   nams <- do.call(c, lapply(x, function(i) i@name))
   dat_obs <- x[[1]]$data$observed
   not_num <- !sapply(dat_obs, inherits, what = "numeric")
@@ -86,28 +87,6 @@ pmc_srmr.mixture_list <- function(x, ..., reps = 100, ci = .95){
   return(out)
 }
 
-# #' @importFrom nonnest2 llcont
-# ic_confint <- function(x, y, conf.level = .95){
-#   ll1 <- nonnest2::llcont(x)
-#   ll2 <- nonnest2::llcont(y)
-#   n <- x$data$numObs
-#   omega.hat.2 <- (n - 1)/n * var(ll1 - ll2, na.rm = TRUE)
-#   bic1 <- AIC(object1, k = log(n))
-#   aic1 <- AIC(object1)
-#   bic2 <- AIC(object2, k = log(n))
-#   aic2 <- AIC(object2)
-#   bicdiff <- bic1 - bic2
-#   aicdiff <- aic1 - aic2
-#   alpha <- 1 - conf.level
-#   BICci <- bicdiff + qnorm(c(alpha/2, (1 - alpha/2))) * sqrt(n *
-#                                                                4 * omega.hat.2)
-#   AICci <- aicdiff + qnorm(c(alpha/2, (1 - alpha/2))) * sqrt(n *
-#                                                                4 * omega.hat.2)
-#   out <- c(bic1, bic2, BICci, aic1, aic2, AICci)
-#   names(out) <- c(paste0("BIC", 1:2), paste0("dBIC_", c("lb", "ub")), paste0("AIC", 1:2), paste0("dAIC_", c("lb", "ub")))
-#   return(out)
-# }
-
 
 #' @title Calculate Standardized Root Mean Residual
 #' @description Given two datasets, computes the correlation matrix for both,
@@ -134,19 +113,53 @@ srmr <- function(x, y){
   sqrt(mean(dif)^2)
 }
 
-
-#' @rdname pmc_srmr
-#' @param FUN Function used to compare the real data (referred to as `x`) to the
-#' model-implied data (referred to as `y`).
+#' @title Calculate Chi Square Statistic
+#' @description Given two datasets with ordinal variables,
+#' computes the lambda statistic, which is the chi-square statistic minus the
+#' degrees of freedom.
+#' @param x An object for which a method exists
+#' (e.g, `data.frame`).
+#' @param y An object for which a method exists
+#' (e.g, `data.frame`).
+#' @return `numeric`
+#' @examples
+#' \dontrun{
+#' if(interactive()){
+#'  chi_sq(iris[1:2], iris[3:4])
+#' }
+#' }
+#' @rdname chi_sq
 #' @export
-pmc <- function(x, ..., reps = 20, ci = .95, FUN = srmr(x, y)){
+chi_sq <- function(x, y){
+  tab_obs  <- table(x)
+  tab_sim <- table(y)
+  term <- tab_obs*log(tab_obs/tab_sim)
+  term[is.infinite(term) | is.nan(term) | is.na(term)] <- 0
+  2*sum(term)
+}
+
+#' @rdname pmc
+#' @param FUN Function used to compare the real data (referred to as `x`) to the
+#' model-implied data (referred to as `y`). Defaults to `NULL`, which uses
+#' [tidySEM::chi_sq()] for models with all ordinal variables, and
+#' [tidySEM::srmr()] otherwise, treating all variables as continuous.
+#' @export
+pmc <- function(x, ..., reps = 20, ci = .95, FUN = NULL){
   UseMethod("pmc", x)
 }
 
 #' @method pmc mixture_list
 #' @export
-pmc.mixture_list <- function(x, ..., reps = 100, ci = .95, FUN = srmr(x, y)){
-  FUN <- match.fun(FUN)
+pmc.mixture_list <- function(x, ..., reps = 100, ci = .95, FUN = NULL){
+  FUN <- try(match.fun(FUN), silent = TRUE)
+  if(inherits(FUN, what = "try-error")){
+    if(all(sapply(x[[1]]$data$observed, inherits, what = "ordered"))){
+      FUN <- chi_sq
+    } else {
+      message("No argument `FUN` provided, resorting to default function `srmr()` and treating all variables as numeric.")
+      return(pmc_srmr_mixture_list(x = x, reps = reps, ci = ci))
+    }
+  }
   nams <- do.call(c, lapply(x, function(i) i@name))
   eval_env <- new.env()
   assign("x", x[[1]]$data$observed, envir = eval_env)
@@ -163,8 +176,7 @@ pmc.mixture_list <- function(x, ..., reps = 100, ci = .95, FUN = srmr(x, y)){
           expr = {
             pgs(sprintf(progmsg))
             assign("y", OpenMx::mxGenerateData(x[[i]]), envir = eval_env)
-            browser()
-            as.numeric(eval(FUN, envir = eval_env))
+            as.numeric(eval(body(FUN), envir = eval_env))
           }
         ))
       }))
@@ -178,7 +190,6 @@ pmc.mixture_list <- function(x, ..., reps = 100, ci = .95, FUN = srmr(x, y)){
   }
   meds <- apply(rep_stat, 2, median)
   names(meds) <- nams
-
   cis <- list(
     dif_seq = rep_stat[, 2:dim(rep_stat)[2]] - rep_stat[, 1:(dim(rep_stat)-1L)[2]],
     dif_one = rep_stat[, 2:dim(rep_stat)[2]] - rep_stat[, rep(1, (dim(rep_stat)[2]-1L))]
@@ -190,8 +201,8 @@ pmc.mixture_list <- function(x, ..., reps = 100, ci = .95, FUN = srmr(x, y)){
     alt = rep(nams[2:length(nams)], 2))
   out <- data.frame(comparison = rep(c("dif_seq", "dif_one"), each = nrow(out)/2),
                     out)
-  out$null <- meds[out$null]
-  out$alt <- meds[out$alt]
+  out$null_stat <- meds[out$null]
+  out$alt_stat <- meds[out$alt]
   out <- data.frame(out, cis, sig = c("", "*")[(apply(sign(cis), 1, sum) == -2)+1L])
   return(out)
 }